Rename output-area -> outputarea
