+++ src/cells/widget.ts
@@ -35,7 +35,7 @@ import {
 
 import {
   OutputAreaWidget
-} from '../output-area';
+} from '../outputarea';
 
 import {
   ICellModel, ICodeCellModel,
