+++ src/console/panel.ts
@@ -39,7 +39,7 @@ import {
 
 import {
   OutputAreaWidget
-} from '../output-area';
+} from '../outputarea';
 
 import {
   IRenderMime
