+++ src/console/widget.ts
@@ -37,7 +37,7 @@ import {
 
 import {
   OutputAreaWidget
-} from '../output-area';
+} from '../outputarea';
 
 import {
   IRenderMime
