+++ src/default-theme/index.css
@@ -33,7 +33,7 @@
 @import '../launcher/index.css';
 @import '../markdownwidget/index.css';
 @import '../notebook/index.css';
-@import '../output-area/index.css';
+@import '../outputarea/index.css';
 @import '../renderers/index.css';
 @import '../running/index.css';
 @import '../terminal/index.css';
