+++ src/notebook/notebook/widget.ts
@@ -77,7 +77,7 @@ import {
 
 import {
   OutputAreaWidget
-} from '../../output-area';
+} from '../../outputarea';
 
 import {
   INotebookModel
