+++ src/outputarea/index.css
