+++ src/outputarea/index.ts
