+++ src/outputarea/model.ts
