+++ src/outputarea/widget.ts
